Add Nextcloud docs link to OPcache recommends
A link to the Nextcloud documentation is currently only shown when the OPcache module is not loaded at all. This commit moves the link to the generic text above the individual recommendations list.

Additionally remove the obsolete phpOpcacheDocumentation entry from test data arrays, which is not passed anymore by the backend.

Signed-off-by: MichaIng <micha@dietpi.com>